+++ lib/index.ts
@@ -1,7 +1,12 @@
 import { RushObserver } from "./observer/rush-observer";
 import { RushStream } from "./stream/rush-stream";
 import { RushSubscriber } from "./stream/rush-subscriber";
-import { streamFromEvent, streamFromTarget } from "./utils/from-events";
+import {
+  streamFromDynamicTargets,
+  streamFromEvent,
+  streamFromEvents,
+  streamFromTarget, streamFromTargets
+} from "./utils/from-events";
 
 import type {
   RushDebugHook,
@@ -13,8 +18,7 @@ import type {
 
 
 export {
-  RushObserver, RushStream, RushSubscriber,
-  streamFromEvent, streamFromTarget
+  RushObserver, RushStream, RushSubscriber, streamFromDynamicTargets, streamFromEvent, streamFromEvents, streamFromTarget, streamFromTargets
 };
 
   export type {